refactor: move DB writes to OrderPersistenceService to activate Transactional via proxy
