+++ src/main/kotlin/ecommerce/service/OrderService.kt
@@ -3,13 +3,7 @@ package ecommerce.service
 import ecommerce.dto.PaymentRequest
 import ecommerce.dto.PlaceOrderRequest
 import ecommerce.dto.PlaceOrderResponse
-import ecommerce.dto.StripeIntentResponse
 import ecommerce.entity.Member
-import ecommerce.entity.Order
-import ecommerce.entity.OrderItem
-import ecommerce.entity.Payment
-import ecommerce.enums.OrderStatus
-import ecommerce.enums.PaymentMethod
 import ecommerce.infrastructure.StripeClient
 import ecommerce.repository.CartItemRepositoryJpa
 import ecommerce.repository.CartRepositoryJpa
@@ -18,9 +12,7 @@ import ecommerce.repository.OrderItemRepositoryJpa
 import ecommerce.repository.OrderRepositoryJpa
 import ecommerce.repository.PaymentRepositoryJpa
 import ecommerce.util.MoneyUtil
-import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
-import java.time.LocalDateTime
 
 @Service
 class OrderService(
@@ -31,6 +23,7 @@ class OrderService(
     private val cartItemRepository: CartItemRepositoryJpa,
     private val cartRepository: CartRepositoryJpa,
     private val stripeClient: StripeClient,
+    private val orderPersistenceService: OrderPersistenceService,
 ) {
     fun place(
         member: Member,
@@ -54,7 +47,7 @@ class OrderService(
         }
 
         val orderId =
-            persistAfterStripeSuccess(
+            orderPersistenceService.persistAfterStripeSuccess(
                 member,
                 option.id!!,
                 req.quantity,
@@ -64,64 +57,5 @@ class OrderService(
                 req.paymentMethod,
             )
         return PlaceOrderResponse(orderId, stripeRes.status, "Order placed and paid successfully.")
-    }
-
-    @Transactional
-    fun persistAfterStripeSuccess(
-        member: Member,
-        optionId: Long,
-        requestedQty: Long,
-        amountMinor: Long,
-        currency: String,
-        stripeRes: StripeIntentResponse,
-        paymentMethod: PaymentMethod,
-    ): Long {
-        val option =
-            optionRepository.findById(optionId)
-                .orElseThrow { IllegalArgumentException("Option not found during persist: $optionId") }
-        require(option.quantity >= requestedQty) { "Insufficient stock during persist." }
-
-        val order =
-            orderRepository.save(
-                Order(
-                    memberId = member.id ?: error("Member must be persisted (id is null)"),
-                    status = OrderStatus.PAID,
-                    orderDateTime = LocalDateTime.now(),
-                ),
-            )
-
-        orderItemRepository.save(
-            OrderItem(
-                order = order,
-                productOption = option,
-                quantity = requestedQty.toInt(),
-                price = option.product.price,
-                productName = option.product.name,
-                optionName = option.name,
-                productImageUrl = option.product.imageUrl,
-            ),
-        )
-
-        paymentRepository.save(
-            Payment(
-                order = order,
-                amount = amountMinor,
-                currency = currency,
-                status = "PAID",
-                stripeSessionId = stripeRes.id,
-                paymentMethod = paymentMethod,
-            ),
-        )
-
-        option.decreaseQuantity(requestedQty)
-        optionRepository.save(option)
-
-        cartRepository.findByMemberId(member.id!!)?.let { cart ->
-            cartItemRepository.findByCartIdAndProductOptionId(cart.id!!, option.id!!)?.let {
-                cartItemRepository.delete(it)
-            }
-        }
-
-        return order.id!!
     }
 }
